+++ core/types/snapshot_nodes.go
@@ -3478,13 +3478,25 @@ func ExecMarketFromProto(em *snapshot.Market) *ExecMarket {
 		StopOrders:                 em.StopOrders,
 		Product:                    em.Product,
 	}
+
 	for _, o := range em.ExpiringOrders {
 		or, _ := OrderFromProto(o)
 		ret.ExpiringOrders = append(ret.ExpiringOrders, or)
 	}
-	for _, o := range em.ExpiringStopOrders {
-		ret.ExpiringStopOrders = append(ret.ExpiringOrders, &Order{ID: o.Id, ExpiresAt: o.ExpiresAt})
+
+	// TODO: remove this on the next minor
+	if em.StopOrders != nil {
+		for _, v := range em.StopOrders.StopOrders {
+			if v.StopOrder.ExpiresAt != nil {
+				ret.ExpiringStopOrders = append(ret.ExpiringStopOrders, &Order{ID: v.StopOrder.Id, ExpiresAt: *v.StopOrder.ExpiresAt})
+			}
+		}
 	}
+
+	// TODO: enable this then
+	// for _, o := range em.ExpiringStopOrders {
+	//     ret.ExpiringStopOrders = append(ret.ExpiringStopOrders, &Order{ID: o.Id, ExpiresAt: o.ExpiresAt})
+	// }
 	return &ret
 }
 
@@ -3527,7 +3539,7 @@ func (e ExecMarket) IntoProto() *snapshot.Market {
 		ret.ExpiringOrders = append(ret.ExpiringOrders, o.IntoProto())
 	}
 	for _, o := range e.ExpiringStopOrders {
-		ret.ExpiringStopOrders = append(ret.ExpiringOrders, &vega.Order{Id: o.ID, ExpiresAt: o.ExpiresAt})
+		ret.ExpiringStopOrders = append(ret.ExpiringStopOrders, &vega.Order{Id: o.ID, ExpiresAt: o.ExpiresAt})
 	}
 	return &ret
 }