+++ CHANGELOG.md
@@ -126,6 +126,7 @@
 - [9153](https://github.com/vegaprotocol/vega/issues/9153) - `MTM` win transfers can be less than one.
 - [9178](https://github.com/vegaprotocol/vega/issues/9178) - Fix LP amendment panic
 - [9053](https://github.com/vegaprotocol/vega/issues/9053) - Handle settle market events in core positions plug-in.
+- [9189](https://github.com/vegaprotocol/vega/issues/9189) - Fix stop orders snapshots
 
 ## 0.72.1